fix(steering): fixed issue with boid orientation matching
The velocity.y is not supposed to be negative since the velocity is based on the frame of the boid so we can assume a x-right, y-up convention. Therefore we don't need to correct for the convention. The 90 deg correction is not necessary since the boid image is facing the right so it faces the x axis. Otherwise we have to correct for the orientation of the image.